requirements.txt
# Core dependencies
pydantic-ai>=0.0.2
pydantic>=2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# CLI
click>=8.1.0

# Async support
aiohttp>=3.9.0
httpx>=0.25.0

# Image generation APIs
openai>=1.0.0
replicate>=0.25.0
stability-sdk>=0.1.0
huggingface-hub>=0.20.0
runwayml>=0.1.0
leonardo-ai-sdk>=1.0.0

# Logging
colorama>=0.4.6